+++ tests/testthat/test-import.R
@@ -1,0 +1,6 @@
+test_that("import() works", {
+  import("tools", "file_ext")
+  expect_identical(file_ext, tools::file_ext)
+  expect_error(import("tools", "file_ext"))
+  expect_error(import("tools", "file_ext", overwrite = TRUE), NA)
+})