+++ NEWS.md
@@ -15,6 +15,7 @@
 * `flip.matrix(, keep_rownames = FALSE)` now works correctly
 * `any_match()` now works as expected
 * `lines_of_r_code()` now works correctly reading a single file
+* `import(, overwrite = TRUE)` now works
 
 ## Non visible changes